@@ -11,7 +11,9 @@ ignore_missing_imports = true
 
 [tool.ruff]
 line-length = 88
-select = [
+target-version = "py39"
+
+lint.select = [
     # pyflakes
     "F",
     # pycodestyle
@@ -52,10 +54,11 @@ select = [
     "UP",
     # pydocstyle
     "D",
+    # isort
+    "I",
 ]
 
-target-version = "py38"
-ignore = [ # space before : (needed for how black formats slicing)
+lint.ignore = [ # space before : (needed for how black formats slicing)
     # "E203",  # not yet implemented
     # do not assign a lambda expression, use a def
     "E731",
@@ -141,12 +144,10 @@ ignore = [ # space before : (needed for how black formats slicing)
 exclude = [
     "docs/*",
     "local_ignore/*",
-    "versioneer.py",
-    "geopandas/_version.py",
 ]
 
-[tool.ruff.per-file-ignores]
+[tool.ruff.lint.per-file-ignores]
 "tests/*" = ["D"]
 
-[tool.ruff.pydocstyle]
+[tool.ruff.lint.pydocstyle]
 convention = "google"
