@@ -6,9 +6,9 @@ import datetime
 import importlib
 import inspect
 import logging
-from pathlib import Path
 import tempfile
-from typing import List, Optional
+from pathlib import Path
+from typing import Optional
 
 import pandas as pd
 
@@ -65,8 +65,8 @@ class RunResult:
 
 
 def run_benchmarks(
-    modules_to_run: Optional[List[str]] = None,
-    functions_to_run: Optional[List[str]] = None,
+    modules_to_run: Optional[list[str]] = None,
+    functions_to_run: Optional[list[str]] = None,
 ):
     """
     Run all benchmarks specified.
