@@ -3,45 +3,44 @@ Module with helper functions for geo files.
 """
 
 import enum
-from datetime import date, datetime
 import filecmp
 import logging
-from pathlib import Path
 import pprint
 import shutil
 import string
 import tempfile
 import time
+import warnings
+from collections.abc import Iterable
+from datetime import date, datetime
+from pathlib import Path
 from typing import (
+    TYPE_CHECKING,
     Any,
-    Dict,
-    Iterable,
-    List,
     Literal,
     Optional,
-    Tuple,
     Union,
-    TYPE_CHECKING,
 )
-import warnings
 
 import geopandas as gpd
-from geopandas.io import file as gpd_io_file
 import numpy as np
-from osgeo import gdal
 import pandas as pd
-from pandas.api.types import is_integer_dtype
-from pygeoops import GeometryType, PrimitiveType  # noqa: F401
 import pyogrio
 import pyproj
+from geopandas.io import file as gpd_io_file
+from osgeo import gdal
+from pandas.api.types import is_integer_dtype
+from pygeoops import GeometryType, PrimitiveType  # noqa: F401
 
-from geofileops.util import _geofileinfo
-from geofileops.util import _geoseries_util
-from geofileops.util import _io_util
-from geofileops.util import _ogr_util
-from geofileops.util import _ogr_sql_util
 from geofileops._compat import PYOGRIO_GTE_07
 from geofileops.helpers._configoptions_helper import ConfigOptions
+from geofileops.util import (
+    _geofileinfo,
+    _geoseries_util,
+    _io_util,
+    _ogr_sql_util,
+    _ogr_util,
+)
 
 if TYPE_CHECKING:
     import os
@@ -106,7 +105,7 @@ PRJ_EPSG_31370 = (
 def listlayers(
     path: Union[str, "os.PathLike[Any]"],
     only_spatial_layers: bool = True,
-) -> List[str]:
+) -> list[str]:
     """
     Get the list of layers in a geofile.
 
@@ -210,14 +209,14 @@ class LayerInfo:
         self,
         name: str,
         featurecount: int,
-        total_bounds: Tuple[float, float, float, float],
+        total_bounds: tuple[float, float, float, float],
         geometrycolumn: str,
         geometrytypename: str,
         geometrytype: GeometryType,
-        columns: Dict[str, ColumnInfo],
+        columns: dict[str, ColumnInfo],
         fid_column: str,
         crs: Optional[pyproj.CRS],
-        errors: List[str],
+        errors: list[str],
     ):
         """
         Constructor of Layerinfo.
@@ -252,7 +251,7 @@ class LayerInfo:
 
 def get_layer_geometrytypes(
     path: Union[str, "os.PathLike[Any]"], layer: Optional[str] = None
-) -> List[str]:
+) -> list[str]:
     """
     Get the geometry types in the layer by examining each geometry in the layer.
 
@@ -1395,7 +1394,7 @@ def _fill_out_sql_placeholders(
     ]
     layer_tmp = layer
     layerinfo = None
-    format_kwargs: Dict[str, Any] = {}
+    format_kwargs: dict[str, Any] = {}
     for placeholder in placeholders:
         if layer_tmp is None:
             layer_tmp = get_only_layer(path)
@@ -2583,7 +2582,7 @@ def copy_layer(
     )
 
 
-def _launder_column_names(columns: Iterable) -> List[Tuple[str, str]]:
+def _launder_column_names(columns: Iterable) -> list[tuple[str, str]]:
     """
     Launders the column names passed to comply with shapefile restrictions.
 
