@@ -3,21 +3,21 @@ Module containing utilities regarding operations on geoseries.
 """
 
 import logging
-from typing import List, Union
 import warnings
+from typing import Union
 
 import geopandas as gpd
 import geopandas._compat as gpd_compat
 import numpy as np
-from numpy.typing import NDArray
 import pandas as pd
-from pygeoops import GeometryType
-from pygeoops._general import _extract_0dim_ndarray
 import pygeoops
 import shapely
+from numpy.typing import NDArray
+from pygeoops import GeometryType
+from pygeoops._general import _extract_0dim_ndarray
 from shapely.geometry.base import BaseGeometry
 
-if gpd_compat.USE_PYGEOS:
+if hasattr(gpd_compat, "USE_PYGEOS") and gpd_compat.USE_PYGEOS:
     import pygeos as shapely2_or_pygeos
 else:
     import shapely as shapely2_or_pygeos
@@ -28,7 +28,7 @@ logger = logging.getLogger(__name__)
 
 def get_geometrytypes(
     geoseries: gpd.GeoSeries, ignore_empty_geometries: bool = True
-) -> List[GeometryType]:
+) -> list[GeometryType]:
     """
     Determine the geometry types in the GeoDataFrame.
 
@@ -117,7 +117,7 @@ def _harmonize_to_multitype(
     geoseries: gpd.GeoSeries, dest_geometrytype: GeometryType
 ) -> gpd.GeoSeries:
     # Copy geoseries data to new array
-    if gpd_compat.USE_PYGEOS:
+    if hasattr(gpd_compat, "USE_PYGEOS") and gpd_compat.USE_PYGEOS:
         geometries_arr = geoseries.array.data.copy()
     else:
         geometries_arr = geoseries.copy()
