@@ -2,7 +2,8 @@
 Module with utilities to format sql statements meant for use with ogr.
 """
 
-from typing import Iterable, List, Optional
+from collections.abc import Iterable
+from typing import Optional
 
 
 class ColumnFormatter:
@@ -13,11 +14,11 @@ class ColumnFormatter:
     regarding the handling of the special "fid" column.
     """
 
-    _aliases_cache: Optional[List[str]] = None
+    _aliases_cache: Optional[list[str]] = None
 
     def __init__(
         self,
-        columns_asked: Optional[List[str]],
+        columns_asked: Optional[list[str]],
         columns_in_layer: Iterable[str],
         fid_column: str,
         table_alias: str = "",
@@ -79,14 +80,14 @@ class ColumnFormatter:
         self._columns = columns
         self._columns_asked = columns_asked
 
-    def _columns_prefixed(self) -> List[str]:
+    def _columns_prefixed(self) -> list[str]:
         columns_prefixed = [
             f'{self._table_prefix}"{column}"' for column in self._columns
         ]
         columns_prefixed = self._fix_fid_columns(columns_prefixed)
         return columns_prefixed
 
-    def _fix_fid_columns(self, columns: List[str]) -> List[str]:
+    def _fix_fid_columns(self, columns: list[str]) -> list[str]:
         """
         Fix the fid columns.
 
@@ -121,7 +122,7 @@ class ColumnFormatter:
 
         return columns
 
-    def _aliases(self) -> List[str]:
+    def _aliases(self) -> list[str]:
         if self._aliases_cache is not None:
             return self._aliases_cache
 
@@ -185,7 +186,7 @@ class ColumnFormatter:
         return f",{', '.join(columns_from_subselect)}"
 
 
-def columns_quoted(columns: List[str]):
+def columns_quoted(columns: list[str]):
     if len(columns) == 0:
         return ""
     columns_quoted = [f'"{column}"' for column in columns]
