@@ -4,13 +4,13 @@ Module containing utilities regarding the usage of ogr/gdal functionalities.
 
 import logging
 import os
-from pathlib import Path
 import tempfile
+from collections.abc import Iterable
+from pathlib import Path
 from threading import Lock
-from typing import Dict, Iterable, List, Literal, Optional, Tuple, Union
+from typing import Literal, Optional, Union
 
-from osgeo import gdal
-from osgeo import ogr
+from osgeo import gdal, ogr
 from pygeoops import GeometryType
 
 import geofileops as gfo
@@ -29,8 +29,8 @@ class GDALError(Exception):
     def __init__(
         self,
         message: str,
-        log_details: List[str] = [],
-        error_details: List[str] = [],
+        log_details: list[str] = [],
+        error_details: list[str] = [],
     ):
         self.message = message
         self.log_details = log_details
@@ -110,7 +110,7 @@ def get_drivers() -> dict:
     return drivers
 
 
-def read_cpl_log(path: Path) -> Tuple[List[str], List[str]]:
+def read_cpl_log(path: Path) -> tuple[list[str], list[str]]:
     """
     Reads a cpl_log file and returns a list with log lines and errors.
 
@@ -145,13 +145,13 @@ class VectorTranslateInfo:
         self,
         input_path: Path,
         output_path: Path,
-        input_layers: Union[List[str], str, None] = None,
+        input_layers: Union[list[str], str, None] = None,
         output_layer: Optional[str] = None,
         input_srs: Union[int, str, None] = None,
         output_srs: Union[int, str, None] = None,
         reproject: bool = False,
-        spatial_filter: Optional[Tuple[float, float, float, float]] = None,
-        clip_geometry: Optional[Union[Tuple[float, float, float, float], str]] = None,
+        spatial_filter: Optional[tuple[float, float, float, float]] = None,
+        clip_geometry: Optional[Union[tuple[float, float, float, float], str]] = None,
         sql_stmt: Optional[str] = None,
         sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
         where: Optional[str] = None,
@@ -220,13 +220,13 @@ def vector_translate_by_info(info: VectorTranslateInfo):
 def vector_translate(
     input_path: Union[Path, str],
     output_path: Path,
-    input_layers: Union[List[str], str, None] = None,
+    input_layers: Union[list[str], str, None] = None,
     output_layer: Optional[str] = None,
     input_srs: Union[int, str, None] = None,
     output_srs: Union[int, str, None] = None,
     reproject: bool = False,
-    spatial_filter: Optional[Tuple[float, float, float, float]] = None,
-    clip_geometry: Optional[Union[Tuple[float, float, float, float], str]] = None,
+    spatial_filter: Optional[tuple[float, float, float, float]] = None,
+    clip_geometry: Optional[Union[tuple[float, float, float, float], str]] = None,
     sql_stmt: Optional[str] = None,
     sql_dialect: Optional[Literal["SQLITE", "OGRSQL"]] = None,
     where: Optional[str] = None,
@@ -266,10 +266,6 @@ def vector_translate(
         input_srs = f"EPSG:{input_srs}"
 
     # Sql'ing, Filtering, clipping
-    if spatial_filter is not None:
-        args.extend(["-spat"])
-        bounds = [str(coord) for coord in spatial_filter]
-        args.extend(bounds)
     if sql_stmt is not None:
         # If sql_stmt starts with "\n" or "\t" for gpkg or with " " for a shp,
         # VectorTranslate outputs no or an invalid file if the statement doesn't return
@@ -621,7 +617,7 @@ def _prepare_gdal_options(options: dict, split_by_option_type: bool = False) -> 
         "DESTINATION_OPEN",
         "CONFIG",
     ]
-    prepared_options: Dict[str, dict] = {
+    prepared_options: dict[str, dict] = {
         option_type: {} for option_type in option_types
     }
 
@@ -687,6 +683,6 @@ class set_config_options:
     def __exit__(self, type, value, traceback):
         # Remove config options that were set
         # TODO: delete loop + uncomment if SetConfigOptions() is supported
-        for name, value in self.config_options.items():
+        for name, _ in self.config_options.items():
             gdal.SetConfigOption(name, None)
         # gdal.SetConfigOptions(self.config_options_backup)
