@@ -2,7 +2,8 @@
 Tests for functionalities in sql_util.
 """
 
-from typing import Iterable, List, Optional
+from collections.abc import Iterable
+from typing import Optional
 
 import pytest
 
@@ -85,7 +86,7 @@ from geofileops.util import _ogr_sql_util
 )
 def test_ColumnFormatter(
     descr: str,
-    columns_specified: Optional[List[str]],
+    columns_specified: Optional[list[str]],
     columns_available: Iterable[str],
     table_alias: str,
     columnname_prefix: str,
